% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignerPeruvianGrid.R
\name{assignerPeruvianGrid}
\alias{assignerPeruvianGrid}
\title{Title}
\usage{
assignerPeruvianGrid(
  data,
  xlon = "lon",
  ylat = "lat",
  by = "onedegree",
  metadata = F,
  save = T,
  cout = "."
)
}
\arguments{
\item{cout}{}
}
\description{
Title
}
