% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{curve.sel}
\alias{curve.sel}
\title{Title}
\usage{
curve.sel(
  sp = "anc",
  stk = 1,
  par = c(NA, NA),
  method = "log3",
  add.plot = T,
  add.inv = T
)
}
\arguments{
\item{add.inv}{}
}
\description{
Title
}
