% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_sel.r
\name{curve_sel}
\alias{curve_sel}
\title{Title}
\usage{
curve_sel(
  sp = NA,
  stock = NA,
  par = c(NA, NA),
  method = "log3",
  add.plot = T,
  add.inv = T
)
}
\arguments{
\item{add.inv}{}
}
\description{
Title
}
