% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateMode.R
\name{estimateMode}
\alias{estimateMode}
\title{Estima moda de una frecuencia de tallas}
\usage{
estimateMode(len, freq, tol.n = 150, nmodes = 2, tol.freq = 10)
}
\arguments{
\item{len}{vector completo de tallas}

\item{freq}{vector completo de frecuencias}

\item{tol.n}{valor de tolerancia del numero de la muestra}

\item{nmodes}{numero de modas a estimar}

\item{tol.freq}{valor de tolerancia para ser considerada moda}
}
\description{
Estima moda de una frecuencia de tallas
}
