% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2plotSurvey.R
\name{g2plotSurvey}
\alias{g2plotSurvey}
\title{Title}
\usage{
g2plotSurvey(
  dat = dat,
  col.var = "biomasa",
  col.factor = 2,
  factor = NA,
  cols.factor = c("red", "blue"),
  box = "boxplot",
  model = "lm",
  lambda = 1e-04,
  IC.model = TRUE,
  info = TRUE,
  marf = c(0.5, 0.5, 0, 0.5),
  omaf = c(3, 4.5, 2, 2),
  hline = TRUE,
  cline = "white",
  bgf = "white",
  lwdf = 1,
  ltyf = 1,
  pchf = 16,
  cexf = 1.5,
  unitf = 1,
  saveplot = T,
  outf = "./",
  widthf = 1800,
  heightf = 1200,
  resf = 250,
  labels.f = c("\\nverano", "\\ninvierno-\\nprimavera"),
  labels.y = "Biomasa ton",
  ...
)
}
\arguments{
\item{labels.y}{}
}
\description{
Title
}
