% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderBiometric.R
\name{get_biometric}
\alias{get_biometric}
\title{Title}
\usage{
get_biometric(
  cin = "rawdata",
  cout = NA,
  file = "Biometria por especies 03032023.xlsx",
  save = T
)
}
\arguments{
\item{save}{}
}
\description{
Title
}
