% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModeSpecies.R
\name{getModeSpecies}
\alias{getModeSpecies}
\title{Title}
\usage{
getModeSpecies(
  Length = Length,
  sp = "anchoveta",
  src = "fsh",
  nmodes = 2,
  tol = 0,
  savePlot = TRUE,
  dirout = "Outputs/"
)
}
\arguments{
\item{dirout}{}
}
\description{
Title
}
