% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPeruGrid.R
\name{MapPeruGrid}
\alias{MapPeruGrid}
\title{Title}
\usage{
MapPeruGrid(
  data,
  colcode = "code",
  colval = "freq",
  typeval = "#",
  xxlim = c(-86, -70),
  yylim = c(-21, -3),
  by = "onedegree",
  gradient = c("yellow", "red"),
  border = NA,
  all.grid = F,
  land.col = "gray90",
  land.border = "gray90",
  legend = T,
  txtleg = "(n)",
  xaxis = 1,
  yaxis = 4,
  portImport = 2,
  save = T,
  cout = "."
)
}
\arguments{
\item{cout}{}
}
\description{
Title
}
