% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2CompSizeTime.R
\name{plot2CompSizeTime}
\alias{plot2CompSizeTime}
\title{Title}
\usage{
plot2CompSizeTime(
  dat = dat_marks,
  sp = "anchoveta",
  src = "fsh",
  type = "B",
  ab = NA,
  factor = NA,
  cexf = 3,
  col.in = "gray90",
  col.bd = "black",
  marp = c(3, 3, 2, 2),
  omap = c(1, 2, 1, 1),
  mgpp = c(1, 0.5, 0),
  widthFig = 2600,
  heightFig = 3200,
  resFig = 380,
  SavePlot = T,
  dirout = "Outputs/"
)
}
\arguments{
\item{dirout}{}
}
\description{
Title
}
