% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_envir.R
\name{plot_envir}
\alias{plot_envir}
\title{Title}
\usage{
plot_envir(
  what = NA,
  year.limit = c(1990, 2022),
  ylim = c(-2, 3.5),
  magnitude = F,
  axis.x = T
)
}
\arguments{
\item{axis.x}{}
}
\description{
Title
}
