% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2SimpleFrec.R
\name{PlotSimpleFrec}
\alias{PlotSimpleFrec}
\title{Title}
\usage{
PlotSimpleFrec(
  data = tmp2l,
  sp = ".",
  stock = ".",
  col.sp = "red",
  cout = "outputs",
  format = ".pdf",
  npar = c(5, 8),
  show.warnings = T,
  col.warnings = "gray70",
  n.warnings = 20,
  ylim = c(0, 0.6),
  width = 2625,
  height = 1750
)
}
\arguments{
\item{colset}{}
}
\description{
Title
}
