% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSimpleFrec2.R
\name{PlotSimpleFrec2}
\alias{PlotSimpleFrec2}
\title{Title}
\usage{
PlotSimpleFrec2(
  data,
  sp = sp,
  stock = stock,
  col.sp = "red",
  colset = "navajowhite",
  cout = ".",
  save = T,
  format = ".png",
  ylim = c(0, 0.5),
  width = 2625,
  height = 1750
)
}
\arguments{
\item{colset}{}
}
\description{
Title
}
