% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleFrecSets.R
\name{SimpleFrecSets}
\alias{SimpleFrecSets}
\title{Plot: Estructura de tallas para "anchoveta "spTarget" lances}
\usage{
SimpleFrecSets(
  data = tmp,
  sp = "anchoveta",
  stock = "nc",
  col.sp = "red",
  cout = "outputs",
  format = ".pdf",
  npar = c(5, 8),
  show.warnings = T,
  col.warnings = "gray70",
  n.warnings = 20,
  ylim = c(0, 0.6),
  width = 2625,
  height = 1750,
  save = T
)
}
\arguments{
\item{data}{datos del crucero en formato MF}

\item{sp}{especie objetivo de la evaluación}

\item{stock}{stock objetivo de la evaluación}

\item{col.sp}{color de la estructura de tallas en el plot matriz}

\item{cout}{vector de caracteres que contienen la ruta de salida}

\item{format}{formato de salida para el plot matriz}

\item{npar}{dos valores numéricos que definen el numero de filas y columnas
para plot matriz "c(x,y)"}

\item{show.warnings}{valor logico; ¿Se deben resaltar los lances con
tamaño de muestreo menor al n.warnings?}

\item{col.warnings}{color para resaltar los lances warnings}

\item{n.warnings}{valor numérico que define el umbral para los lances warnings}

\item{ylim}{dos valores numéricos, que especifican el límite en el eje y}

\item{save}{valor logico; ¿Se deben guardar la matriz de lances en "csv"?}
}
\description{
Plot: Estructura de tallas para "anchoveta "spTarget" lances
}
