% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleFrecSp.R
\name{SimpleFrecSp}
\alias{SimpleFrecSp}
\title{Plot: Estructura de tallas para "anchoveta "spTarget" lances}
\usage{
SimpleFrecSp(data = x, sp = "anchoveta", stock = "nc", save = T)
}
\arguments{
\item{data}{datos del crucero en formato MF}

\item{sp}{especie objetivo de la evaluación}

\item{stock}{stock objetivo de la evaluación}

\item{save}{valor logico; ¿Se deben guardar la matriz de lances en "csv"?}
}
\description{
Plot: Estructura de tallas para "anchoveta "spTarget" lances
}
