% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSurvey.R
\name{subsetSurvey}
\alias{subsetSurvey}
\title{Title}
\usage{
subsetSurvey(
  dat = dat,
  col.var = c("biomasa", "abundancia"),
  year.lim = c(1996, 2022),
  col.set = "filter",
  set = "ok",
  col.factor = "season",
  col.label = 2,
  col.date = NA,
  col.month = NA,
  col.sems = "semester",
  col.year = "year"
)
}
\arguments{
\item{col.year}{}
}
\description{
Title
}
