% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSp.R
\name{validateSp}
\alias{validateSp}
\title{Identifica errores y valida la base de datos}
\usage{
validateSp(
  data = data,
  sp = "anchoveta",
  stock = NA,
  cout = "outputs",
  file = "validateSp"
)
}
\arguments{
\item{data}{datos del crucero en formato MF}

\item{sp}{especie objetivo de la evaluación}

\item{stock}{stock objetivo de la evaluación}

\item{cout}{vector de caracteres que contienen la ruta de salida}

\item{file}{ector de caracteres que contienen el nombre del archivo}
}
\description{
Identifica errores y valida la base de datos
}
\examples{
#'
\code{validateSp(data = data, sp = "jurel", stock = NA, cout = "outputs",
file = "document.docx")}

}
