% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcxJA.R
\name{plot.matrixfreq}
\alias{plot.matrixfreq}
\title{Title}
\usage{
\method{plot}{matrixfreq}(
  x,
  relative = T,
  clean = T,
  ylim = c(0, 20),
  yinter = 5,
  juvMarks = 51,
  plotCol = "blue",
  juvCol = "red",
  textx = "Longitud total (cm)",
  type = "barplot"
)
}
\arguments{
\item{type}{}
}
\description{
Title
}
