% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderBiometric.R
\name{renderBiometric}
\alias{renderBiometric}
\title{Crea, unifica archivo biometrico}
\usage{
renderBiometric(cin = "inputs", cout = ".", file = "data.xlsx", save = T)
}
\arguments{
\item{cin}{vector de caracteres que contienen la ruta de entrada}

\item{cout}{vector de caracteres que contienen la ruta de salida}

\item{file}{vector de caracteres que contienen el nombre del archivo}

\item{save}{logico; ¿Se deben guardar el archivo?}
}
\description{
Crea, unifica archivo biometrico
}
\examples{
\code{renderBiometric(cin =  "input",
cout = NA,
file = "Biometria por especies 03032023.xlsx",
save = T)}

}
