% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateSp.R
\name{validateSp}
\alias{validateSp}
\title{Identifica errores y valida la base de datos}
\usage{
validateSp(
  data = data,
  sp = "anchoveta",
  stock = NA,
  cout = "outputs",
  file = "validateSp.html"
)
}
\arguments{
\item{file}{}
}
\description{
Identifica errores y valida la base de datos
}
\examples{
#'
\code{validateSp(data = data, sp = "jurel", stock = NA, cout = "outputs",
file = "document.docx")}

}
